% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stability-metrics.R
\name{jaccard_index}
\alias{jaccard_index}
\title{Jaccard index of two significant-pathway sets}
\usage{
jaccard_index(sig_a, sig_b)
}
\arguments{
\item{sig_a, sig_b}{character vectors of pathway IDs called significant
in two ID-mapping branches.}
}
\value{
numeric in [0, 1]. Defined as 1 when both sets are empty.
}
\description{
Jaccard index of two significant-pathway sets
}
